Fix some types & prototypes.

+++ c/librcksum/rcksum.h
@@ -57,6 +57,8 @@ zs_blockid* rcksum_needed_block_ranges(const struct rcksum_state* z, int* num, z
 int rcksum_blocks_todo(const struct rcksum_state*);
 
 /* For preparing rcksum control files - in both cases len is the block size. */
-struct rsum __attribute__((pure)) rcksum_calc_rsum_block(const unsigned char* data, int len);
+struct rsum __attribute__((pure))
+rcksum_calc_rsum_block(const unsigned char *data, size_t len);
+
 void rcksum_calc_checksum(unsigned char *c, const unsigned char* data, size_t len);
